refactor(database): keep integrity errors in v2/database
Collapse the database/errors subpackage and stop returning domain.Error
from v2/database helpers so domain can import NoRowFoundError from the
database package without an import cycle.
